Created setup.go
New Device Setup
Great work learning how to use the fmt package in Go! It is time to test your new knowledge on a larger problem!

You just bought a new smart device and are ready to set it up! But you’ll be the one to program this setup and prompt the user for their information.